Move text track util functionst to one module
